% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_name}
\alias{model_name}
\alias{model_description}
\title{Name and description of a model}
\usage{
model_name(model)

model_description(model)
}
\arguments{
\item{model}{A model.}
}
\value{
\code{model_name()}: the short name used in reports, such as
\code{"log_arima_011_011"}; \code{model_description()}: one line about the method.
}
\description{
Name and description of a model
}
\examples{
model_name(model_log(model_airline()))
}
