% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foresightr-package.R
\docType{package}
\name{foresightr-package}
\alias{foresightr}
\alias{foresightr-package}
\title{foresightr: Forecasts Chosen by What Would Have Worked}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Time series forecasting that replays the past before trusting a model: a rolling-origin backtest, Tashman (2000) \doi{10.1016/S0169-2070(00)00065-0}, refits every candidate at each origin, the choice is made by out-of-sample error, and prediction intervals are quantiles of the errors actually observed, for each horizon and for totals of the next periods. Models include seasonal ARIMA by exact maximum likelihood with automatic orders, Hyndman and Khandakar (2008) \doi{10.18637/jss.v027.i03}; exponential smoothing in state space form, Hyndman, Koehler, Snyder and Grose (2002) \doi{10.1016/S0169-2070(01)00110-8}; the Theta method, Assimakopoulos and Nikolopoulos (2000) \doi{10.1016/S0169-2070(00)00066-2}; a trend with changepoints and events in the manner of Taylor and Letham (2018) \doi{10.1080/00031305.2017.1380080}; TBATS, De Livera, Hyndman and Snyder (2011) \doi{10.1198/jasa.2011.tm09771}; seasonal-trend decompositions by LOESS for one or several seasonal periods, Bandara, Hyndman and Bergmeir (2021) \doi{10.48550/arXiv.2107.13462}; the method of Croston (1972) \doi{10.1057/jors.1972.50} for intermittent demand; and ensembles. The computations are done by the 'Rust' crate 'foresight', bundled with the package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StrategicProjects/foresightr}
  \item \url{https://strategicprojects.github.io/foresightr/}
  \item Report bugs at \url{https://github.com/StrategicProjects/foresightr/issues}
}

}
\author{
\strong{Maintainer}: André Leite \email{leite@castlab.org} (\href{https://orcid.org/0000-0002-4718-9766}{ORCID})

Authors:
\itemize{
  \item André Leite \email{leite@castlab.org} (\href{https://orcid.org/0000-0002-4718-9766}{ORCID})
  \item Marcos Wasiliew \email{marcos.wasiliew@gmail.com} (\href{https://orcid.org/0009-0004-4694-3159}{ORCID})
  \item Hugo Vasconcelos \email{hugo.vasconcelos@ufpe.br} (\href{https://orcid.org/0000-0001-6249-0920}{ORCID})
  \item Carlos Amorim \email{carlos.agaf@ufpe.br} (\href{https://orcid.org/0000-0001-6315-8305}{ORCID})
  \item Diogo Bezerra \email{diogo.bezerra@ufpe.br} (\href{https://orcid.org/0000-0002-1216-8674}{ORCID})
}

Other contributors:
\itemize{
  \item The authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor]
}

}
\keyword{internal}
